smart-calendar/requirements.txt
fastapi
uvicorn
langgraph>=0.2.34,<0.3.0
langchain 
langchain_openai
langchain-core
openai
sentence-transformers
python-dotenv
openai-whisper
pydantic
python-multipart
langgraph_swarm
open_deep_research
loguru


